+++ src/textual/widgets/_tree.py
@@ -207,7 +207,7 @@ class TreeNode(Generic[TreeDataType]):
         """
         self._expanded = True
         self._updates += 1
-        self._tree.post_message(Tree.NodeExpanded(self._tree, self))
+        self._tree.post_message(Tree.NodeExpanded(self))
         if expand_all:
             for child in self.children:
                 child._expand(expand_all)
@@ -240,7 +240,7 @@ class TreeNode(Generic[TreeDataType]):
         """
         self._expanded = False
         self._updates += 1
-        self._tree.post_message(Tree.NodeCollapsed(self._tree, self))
+        self._tree.post_message(Tree.NodeCollapsed(self))
         if collapse_all:
             for child in self.children:
                 child._collapse(collapse_all)
@@ -514,23 +514,15 @@ class Tree(Generic[TreeDataType], ScrollView, can_focus=True):
         parent node in the DOM.
         """
 
-        def __init__(
-            self, tree: Tree[EventTreeDataType], node: TreeNode[EventTreeDataType]
-        ) -> None:
-            self.tree = tree
-            """The tree that sent the message."""
+        def __init__(self, node: TreeNode[EventTreeDataType]) -> None:
             self.node: TreeNode[EventTreeDataType] = node
             """The node that was collapsed."""
             super().__init__()
 
         @property
         def control(self) -> Tree[EventTreeDataType]:
-            """The tree that sent the message.
-
-            This is an alias for [`NodeCollapsed.tree`][textual.widgets.Tree.NodeCollapsed.tree]
-            and is used by the [`on`][textual.on] decorator.
-            """
-            return self.tree
+            """The tree that sent the message."""
+            return self.node.tree
 
     class NodeExpanded(Generic[EventTreeDataType], Message, bubble=True):
         """Event sent when a node is expanded.
@@ -539,23 +531,15 @@ class Tree(Generic[TreeDataType], ScrollView, can_focus=True):
         parent node in the DOM.
         """
 
-        def __init__(
-            self, tree: Tree[EventTreeDataType], node: TreeNode[EventTreeDataType]
-        ) -> None:
-            self.tree = tree
-            """The tree that sent the message."""
+        def __init__(self, node: TreeNode[EventTreeDataType]) -> None:
             self.node: TreeNode[EventTreeDataType] = node
             """The node that was expanded."""
             super().__init__()
 
         @property
         def control(self) -> Tree[EventTreeDataType]:
-            """The tree that sent the message.
-
-            This is an alias for [`NodeExpanded.tree`][textual.widgets.Tree.NodeExpanded.tree]
-            and is used by the [`on`][textual.on] decorator.
-            """
-            return self.tree
+            """The tree that sent the message."""
+            return self.node.tree
 
     class NodeHighlighted(Generic[EventTreeDataType], Message, bubble=True):
         """Event sent when a node is highlighted.
@@ -564,23 +548,15 @@ class Tree(Generic[TreeDataType], ScrollView, can_focus=True):
         parent node in the DOM.
         """
 
-        def __init__(
-            self, tree: Tree[EventTreeDataType], node: TreeNode[EventTreeDataType]
-        ) -> None:
-            self.tree = tree
-            """The tree that sent the message."""
+        def __init__(self, node: TreeNode[EventTreeDataType]) -> None:
             self.node: TreeNode[EventTreeDataType] = node
             """The node that was highlighted."""
             super().__init__()
 
         @property
         def control(self) -> Tree[EventTreeDataType]:
-            """The tree that sent the message.
-
-            This is an alias for [`NodeHighlighted.tree`][textual.widgets.Tree.NodeHighlighted.tree]
-            and is used by the [`on`][textual.on] decorator.
-            """
-            return self.tree
+            """The tree that sent the message."""
+            return self.node.tree
 
     class NodeSelected(Generic[EventTreeDataType], Message, bubble=True):
         """Event sent when a node is selected.
@@ -589,23 +565,15 @@ class Tree(Generic[TreeDataType], ScrollView, can_focus=True):
         parent node in the DOM.
         """
 
-        def __init__(
-            self, tree: Tree[EventTreeDataType], node: TreeNode[EventTreeDataType]
-        ) -> None:
-            self.tree = tree
-            """The tree that sent the message."""
+        def __init__(self, node: TreeNode[EventTreeDataType]) -> None:
             self.node: TreeNode[EventTreeDataType] = node
             """The node that was selected."""
             super().__init__()
 
         @property
         def control(self) -> Tree[EventTreeDataType]:
-            """The tree that sent the message.
-
-            This is an alias for [`NodeSelected.tree`][textual.widgets.Tree.NodeSelected.tree]
-            and is used by the [`on`][textual.on] decorator.
-            """
-            return self.tree
+            """The tree that sent the message."""
+            return self.node.tree
 
     def __init__(
         self,
@@ -905,7 +873,7 @@ class Tree(Generic[TreeDataType], ScrollView, can_focus=True):
             node._selected = True
             self._cursor_node = node
             if previous_node != node:
-                self.post_message(self.NodeHighlighted(self, node))
+                self.post_message(self.NodeHighlighted(node))
         else:
             self._cursor_node = None
 
@@ -1236,7 +1204,7 @@ class Tree(Generic[TreeDataType], ScrollView, can_focus=True):
 
         Note:
             If `auto_expand` is `True` use of this action on a non-leaf node
-            will cause both an expand/collapse event to occour, as well as a
+            will cause both an expand/collapse event to occur, as well as a
             selected event.
         """
         try:
@@ -1247,4 +1215,4 @@ class Tree(Generic[TreeDataType], ScrollView, can_focus=True):
             node = line.path[-1]
             if self.auto_expand:
                 self._toggle_node(node)
-            self.post_message(self.NodeSelected(self, node))
+            self.post_message(self.NodeSelected(node))